Ability to play error sound instead of long error utterance
==== Fixed Issues ====
issue-2755 -  Ability to play error sound instead of long error utterance

====  Tech Notes ====
Add a new method parameter in the speak method for skills to declare if an utterance is an error.
Add a new error sound. I have no skills in the domain, so it's quite horrible, sorry. It's the wakeword sound, reversed.
Add two new configuration parameters for customizing Mycroft behavior when an error occured (per skill, or global)
Change some TTS as I changed the tts execute signature : they  overrode a method they should not, in my opinion. If my guess is good they should override the inner _execute method and not the main one. I could need some educated guess from the reviewer for this point.
Technical note : The sound is played in the TTS module, instead of the utterance. It could be played upper in the stack but I did this in order to keep the functionality close to its replacement.

====  Documentation Notes ====
Two new parameters can force Mycroft to do error feedback with a brief sound instead of a complete error utterance.
skils.skills_with_sound_as_error : a list of skills who should use sound. Use the class name.
error_as_sound : boolean to force Mycroft to use sound for error feedback, wherever it is possible.

==== Localization Notes ====
NONE

==== Environment Notes ====
NONE

==== Protocol Notes ====
NONE